+++ sdk/eventhub/event-hubs/test/stress/app/src/utils.ts
@@ -3,6 +3,7 @@
 
 import { AbortSignalLike } from "@azure/abort-controller";
 import { delay } from "@azure/core-amqp";
+import { EventHubConsumerClientOptions, EventHubConsumerClient, EventHubClientOptions, EventHubProducerClient } from "@azure/event-hubs";
 
 export interface SnapshotOptions {
   /**
@@ -30,4 +31,33 @@ export async function loopForever(
   while (abortSignal?.aborted === false && (await delay(duration))) {
     await fn();
   }
+}
+
+export function createEventHubsConsumerClient(
+  options?: EventHubConsumerClientOptions
+): EventHubConsumerClient {
+  const consumerGroup = process.env.EVENTHUBS_CONSUMER_GROUP || "$Default";
+  const connectionString = process.env.EVENTHUBS_CONNECTION_STRING;
+  const hubName = process.env.EVENTHUB_NAME;
+  if (!connectionString || !consumerGroup || !hubName) {
+    throw new Error(
+      "EVENTHUBS_CONNECTION_STRING, EVENTHUB_NAME and EVENTHUBS_CONSUMER_GROUP have to be populated in the environment and are not!"
+    );
+  }
+  return new EventHubConsumerClient(consumerGroup, connectionString, hubName, options);
+}
+
+export function createEventHubsProducerClient(
+  options?: EventHubClientOptions
+): EventHubProducerClient {
+  const eventHubName = process.env.EVENTHUB_NAME;
+  const connectionString = process.env.EVENTHUBS_CONNECTION_STRING;
+
+  if (!connectionString || !eventHubName) {
+    throw new Error(
+      "EVENTHUBS_CONNECTION_STRING and EVENTHUB_NAME have to be populated in the environment and are not!"
+    );
+  }
+
+  return new EventHubProducerClient(connectionString, eventHubName, options);
 }
